* 共試驗的用字表改號做教典
