+++ 試驗/教典/test用字語音方言差試驗.py
@@ -1,6 +1,6 @@
 from django.test.testcases import TestCase
 from 臺灣言語工具.基本物件.字 import 字
-from 用字.models import 用字表
+from 用字 import 教典
 
 
 class 用字語音方言差試驗(TestCase):
@@ -9,8 +9,8 @@ class 用字語音方言差試驗(TestCase):
     #    2 1,[方]005,欲,berh,berh,beh,beh,beh,beh,berh,boh,beh,beh
 
     def test_鹿港_粿(self):
-        self.assertTrue(用字表.有這个字無(字('母', 'biór')))
+        self.assertTrue(教典.有這个字無(字('母', 'biór')))
 
     def test_指_無收著臺中腔(self):
         #    26,[方]358,指,tsńg,... ,x
-        self.assertFalse(用字表.有這个字無(字('指', 'x')))
+        self.assertFalse(教典.有這个字無(字('指', 'x')))
